converted all ints/long to int64_t

+++ unionFindLib.h
@@ -5,11 +5,11 @@
 #include <NDMeshStreamer.h>
 
 struct unionFindVertex {
-    long int vertexID;
-    long int parent;
-    long int componentNumber = -1;
-    std::vector<long int> need_boss_requests; //request queue for processing need_boss requests
-    long int findOrAnchorCount = 0;
+    int64_t vertexID;
+    int64_t parent;
+    int64_t componentNumber = -1;
+    std::vector<int64_t> need_boss_requests; //request queue for processing need_boss requests
+    int64_t findOrAnchorCount = 0;
 
     void pup(PUP::er &p) {
         p|vertexID;
@@ -20,8 +20,8 @@ struct unionFindVertex {
 };
 
 struct componentCountMap {
-    long int compNum;
-    int count;
+    int64_t compNum;
+    int64_t count;
 
     void pup(PUP::er &p) {
         p|compNum;
@@ -38,38 +38,38 @@ extern CkReduction::reducerType mergeCountMapsReductionType;
 // class definition for library chares
 class UnionFindLib : public CBase_UnionFindLib {
     std::vector<unionFindVertex> myVertices;
-    int numMyVertices;
-    int pathCompressionThreshold = 5;
+    int64_t numMyVertices;
+    int64_t pathCompressionThreshold = 5;
     int componentPruneThreshold;
-    std::pair<int, int> (*getLocationFromID)(long int vid);
-    int myLocalNumBosses;
-    int totalNumBosses;
+    std::pair<int64_t, int64_t> (*getLocationFromID)(int64_t vid);
+    int64_t myLocalNumBosses;
+    int64_t totalNumBosses;
     CkCallback postComponentLabelingCb;
     
     // Batch
-    long int numCharesinPe;
-    long int batchSize;
-    long int batchNo;
-    long int reqsProcessed;
-    long int totalReqsPerBatch;
-    long int thresholdReqs;
-    long int totalReqsProcessed;
+    int64_t numCharesinPe;
+    int64_t batchSize;
+    int64_t batchNo;
+    int64_t reqsProcessed;
+    int64_t totalReqsPerBatch;
+    int64_t thresholdReqs;
+    int64_t totalReqsProcessed;
     CkCallback batchCb;
 
     public:
     UnionFindLib() {}
     UnionFindLib(CkMigrateMessage *m) { }
-    static CProxy_UnionFindLib unionFindInit(CkArrayID clientArray, int n);
+    static CProxy_UnionFindLib unionFindInit(CkArrayID clientArray, int64_t n);
     void register_phase_one_cb(CkCallback cb);
     // void initialize_vertices(unionFindVertex *appVertices, int numVertices);
-    void allocate_libVertices(long int numVertices, long int nPe);
-    void initialize_vertices(long int numVertices, unionFindVertex* &appVertices, long int &offset, long int bs);
+    void allocate_libVertices(int64_t numVertices, int64_t nPe);
+    void initialize_vertices(int64_t numVertices, unionFindVertex* &appVertices, int64_t &offset, int64_t bs);
 
-    void union_request(long int v, long int w);
-    void anchor(int w_arrIdx, long int v, long int path_base_arrIdx);
-    void local_path_compression(unionFindVertex *src, long int compressedParent);
-    bool check_same_chares(long int v1, long int v2);
-    void registerGetLocationFromID(std::pair<int, int> (*gloc)(long int v));
+    void union_request(int64_t v, int64_t w);
+    void anchor(int64_t w_arrIdx, int64_t v, int64_t path_base_arrIdx);
+    void local_path_compression(unionFindVertex *src, int64_t compressedParent);
+    bool check_same_chares(int64_t v1, int64_t v2);
+    void registerGetLocationFromID(std::pair<int64_t, int64_t> (*gloc)(int64_t v));
     void printVertices();
 
     // functions and data structures for finding connected components
@@ -81,13 +81,13 @@ class UnionFindLib : public CBase_UnionFindLib {
     public:
     void find_components(CkCallback cb);
     void insertDataFindBoss(const findBossData & data);
-    void boss_count_prefix_done(int totalCount);
+    void boss_count_prefix_done(int64_t totalCount);
     void start_component_labeling();
     void insertDataNeedBoss(const uint64_t & data);
     void insertDataAnchor(const anchorData & data);
-    void need_boss(int arrIdx, long int fromID);
-    void set_component(int arrIdx, long int compNum);
-    void prune_components(int threshold, CkCallback appReturnCb);
+    void need_boss(int64_t arrIdx, int64_t fromID);
+    void set_component(int64_t arrIdx, int64_t compNum);
+    void prune_components(int64_t threshold, CkCallback appReturnCb);
     void perform_pruning();
     int get_total_num_bosses() {
         return totalNumBosses;
@@ -95,25 +95,25 @@ class UnionFindLib : public CBase_UnionFindLib {
     //void merge_count_results(CkReductionMsg *msg);
     //void merge_count_results(int* totalCounts, int numElems);
 #ifdef PROFILING
-    void profiling_count_max(long int maxCount);
+    void profiling_count_max(int64_t maxCount);
 #endif
 };
 
 // library group chare class declarations
 class UnionFindLibGroup : public CBase_UnionFindLibGroup {
     bool map_built;
-    int* component_count_array;
-    int thisPeMessages; //for profiling
+    int64_t* component_count_array;
+    int64_t thisPeMessages; //for profiling
     public:
     UnionFindLibGroup() {
         map_built = false;
         thisPeMessages = 0;
     }
-    void build_component_count_array(int* totalCounts, int numComponents);
-    int get_component_count(long int component_id);
+    void build_component_count_array(int64_t* totalCounts, int64_t numComponents);
+    int64_t get_component_count(int64_t component_id);
     void increase_message_count();
     void contribute_count();
-    void done_profiling(int);
+    void done_profiling(int64_t);
 };
 
 
@@ -136,10 +136,10 @@ start_boss_propagation() {
     std::vector<int>::iterator iter = local_boss_indices.begin();
     while (iter != local_boss_indices.end()) {
         int bossIdx = *iter;
-        long int bossID = myVertices[bossIdx].vertexID;
-        std::vector<long int>::iterator req_iter = myVertices[bossIdx].need_boss_requests.begin();
+        int64_t bossID = myVertices[bossIdx].vertexID;
+        std::vector<int64_t>::iterator req_iter = myVertices[bossIdx].need_boss_requests.begin();
         while (req_iter != myVertices[bossIdx].need_boss_requests.end()) {
-            long int requestorID = *req_iter;
+            int64_t requestorID = *req_iter;
             std::pair<int,int> requestor_loc = appPtr->getLocationFromID(requestorID);
             this->thisProxy[requestor_loc.first].set_component(requestor_loc.second, bossID);
             // done with requestor, delete from requests queue
@@ -162,7 +162,7 @@ merge_count_results(CkReductionMsg *msg) {
     }
 
     // convert custom map back to STL, for easier lookup
-    std::map<long int,int> quick_final_map;
+    std::map<int64_t,int> quick_final_map;
     for (int i = 0; i < numComps; i++) {
         if (this->thisIndex == 0) {
             //CkPrintf("Component %ld : Total vertices count = %d\n", final_map[i].compNum, final_map[i].count);
